backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
openai==1.59.7
chromadb==0.6.3
pypdf==5.1.0
